Add the Write method to the xpl module

+++ ansys/mapdl/core/mapdl_grpc.py
@@ -304,7 +304,7 @@ class MapdlGrpc(_MapdlCore):
         request = anskernel.CtrlRequest(ctrl='VERSION')
         resp = self._stub.Ctrl( request)
         grpc_ver = resp.response.split('.')
-        self._grpc_api_ver = float(grpc_ver[0]) + 0.1*float(grpc_ver[1])
+        self._grpc_api_ver = int(grpc_ver[0]) + 0.1*int(grpc_ver[1])
 
         print('\ngRPC api release :    '  + str(self._grpc_api_ver) + '\n')
